requirements_discord.txt
discord.py==2.3.2
anthropic>=0.25.0
supabase>=2.0.0
